refactor : pr 리뷰에 따른 수정
- LectureDocument likes, hits 필드 long -> Integer로 변경
- LectureElasticService searchLectureDocument() like NPE 방지
- LecturePreviewResponseDto Type 필드 NPE 방지
- LectureStartupIndexer
   - totalProcessed 오버플로우 방지
   - index refresh를 통한 가시성 확보
- LectureUploadService 에러 코드 혼재 사용 수정

+++ src/main/java/inu/codin/codin/domain/elasticsearch/document/LectureDocument.java
@@ -52,10 +52,10 @@ public class LectureDocument {
     @Field(type = FieldType.Double)
     private Double starRating;
 
-    @Field(type = FieldType.Long)
+    @Field(type = FieldType.Integer)
     private Integer likes;
 
-    @Field(type = FieldType.Long)
+    @Field(type = FieldType.Integer)
     private Integer hits;
 
     @Field(type = FieldType.Nested)
